fix: refactor therapies tabs into screens
